Start branch for dev/0.6.1

See merge request computer-vision/netharn!22

+++ netharn/util/util_json.py
@@ -144,7 +144,7 @@ def ensure_json_serializable(dict_, normalize_containers=False, verbose=0):
         >>> data['foo'] = ub.ddict(lambda: int)
         >>> data['bar'] = np.array([1, 2, 3])
         >>> data['foo']['a'] = 1
-        >>> data['foo']['b'] = (1, np.array([1, 2, 3]), {3: np.int(3), 4: np.float16(1.0)})
+        >>> data['foo']['b'] = (1, np.array([1, 2, 3]), {3: int(3), 4: np.float16(1.0)})
         >>> dict_ = data
         >>> print(ub.repr2(data, nl=-1))
         >>> result = ensure_json_serializable(data, normalize_containers=True)
